requirements.txt
click==7.1.2
numpy==1.19.4
pygame==2.0.0
PyGLM==1.99.3
PyOpenGL==3.1.5
